requirements.txt
jsonschema==4.6.0
kafka-python==2.0.2
pymongo==4.1.1
python-dotenv==0.20.0